Partial implementation of PIXEL_SelectProfile
Some of the values within TEST_PIXEL_SelectProfile are hard-coded for now.

In the future, a common struct for each profile will be defined, containing common attributes like frame rate, framecount, etc.

In the future, a state variable will be declared which each Profile function can use to determine if there was a recent change in the currentProfileIndex. This way, the function would know when to reset some of its static variables.