menu.c - main menu selection
Wrote a partial implementation of the state-machine-based menu application.

Right now, the test code simply scrolls through the options available in the main menu.

Had to modify test.c to make use of MENU_Tasks().

Certain parts of menu.c are currently "hard-coded" and will need to be updated in the future to make it easier to read.